Revert "Moved the default BinaryMessenger instance to ServicesBinding (#37489)"
This reverts commit 92ef2b9ce1d4e54d5efb3e6a962e9e724dc7efeb.

This requires either runApp() or
WidgetsFlutterBinding.ensureInitialized() to have been called before
using any MethodChannels. Plugins broadly rely on MethodChannels and
right now there's no general requirements that they be constructed
within the runApp call, so the ecosystem breakages from this are broader
than originally thought. Reverting this for now.
