+++ dev/integration_tests/ios_add2app/flutterapp/lib/main.dart
@@ -25,8 +25,6 @@ const BasicMessageChannel<String> _kReloadChannel =
     BasicMessageChannel<String>(_kReloadChannelName, StringCodec());
 
 void main() {
-  // Ensures bindings are initialized before doing anything.
-  WidgetsFlutterBinding.ensureInitialized();
   // Start listening immediately for messages from the iOS side. ObjC calls
   // will be made to let us know when we should be changing the app state.
   _kReloadChannel.setMessageHandler(run);
