+++ packages/flutter/lib/src/widgets/binding.dart
@@ -251,10 +251,6 @@ mixin WidgetsBinding on BindingBase, SchedulerBinding, GestureBinding, RendererB
   void initInstances() {
     super.initInstances();
     _instance = this;
-    // Initialization of [_buildOwner] has to be done after
-    // [super.initInstances] is called, as it requires [ServicesBinding] to
-    // properly setup the [defaultBinaryMessenger] instance.
-    _buildOwner = BuildOwner();
     buildOwner.onBuildScheduled = _handleBuildScheduled;
     window.onLocaleChanged = handleLocaleChanged;
     window.onAccessibilityFeaturesChanged = handleAccessibilityFeaturesChanged;
@@ -375,10 +371,7 @@ mixin WidgetsBinding on BindingBase, SchedulerBinding, GestureBinding, RendererB
   /// The [BuildOwner] in charge of executing the build pipeline for the
   /// widget tree rooted at this binding.
   BuildOwner get buildOwner => _buildOwner;
-  // Initialization of [_buildOwner] has to be done within the [initInstances]
-  // method, as it requires [ServicesBinding] to properly setup the
-  // [defaultBinaryMessenger] instance.
-  BuildOwner _buildOwner;
+  final BuildOwner _buildOwner = BuildOwner();
 
   /// The object in charge of the focus tree.
   ///
