+++ packages/flutter/test/semantics/semantics_service_test.dart
@@ -6,15 +6,10 @@ import 'dart:ui' show TextDirection;
 
 import 'package:flutter/semantics.dart';
 import 'package:flutter/services.dart' show SystemChannels;
-import 'package:flutter_test/flutter_test.dart' show TestWidgetsFlutterBinding;
 
 import '../flutter_test_alternative.dart';
 
 void main() {
-  setUp(() {
-    TestWidgetsFlutterBinding.ensureInitialized();
-  });
-
   test('Semantic announcement', () async {
     final List<Map<dynamic, dynamic>> log = <Map<dynamic, dynamic>>[];
