+++ packages/flutter/test/services/haptic_feedback_test.dart
@@ -6,10 +6,6 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  setUp(() {
-    TestWidgetsFlutterBinding.ensureInitialized();
-  });
-
   test('Haptic feedback control test', () async {
     final List<MethodCall> log = <MethodCall>[];
