+++ packages/flutter/test/services/platform_channel_test.dart
@@ -6,20 +6,14 @@ import 'dart:async';
 import 'dart:typed_data';
 
 import 'package:flutter/services.dart';
-import 'package:flutter_test/flutter_test.dart' show TestWidgetsFlutterBinding;
-
 import '../flutter_test_alternative.dart';
 
 void main() {
-  setUp(() {
-    TestWidgetsFlutterBinding.ensureInitialized();
-  });
-
   group('BasicMessageChannel', () {
     const MessageCodec<String> string = StringCodec();
     const BasicMessageChannel<String> channel = BasicMessageChannel<String>('ch', string);
     test('can send string message and get reply', () async {
-      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
+      defaultBinaryMessenger.setMockMessageHandler(
         'ch',
         (ByteData message) async => string.encodeMessage(string.decodeMessage(message) + ' world'),
       );
@@ -29,7 +23,7 @@ void main() {
     test('can receive string message and send reply', () async {
       channel.setMessageHandler((String message) async => message + ' world');
       String reply;
-      await ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage(
+      await defaultBinaryMessenger.handlePlatformMessage(
         'ch',
         const StringCodec().encodeMessage('hello'),
         (ByteData replyBinary) {
@@ -45,7 +39,7 @@ void main() {
     const MethodCodec jsonMethod = JSONMethodCodec();
     const MethodChannel channel = MethodChannel('ch7', jsonMethod);
     test('can invoke method and get result', () async {
-      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
+      defaultBinaryMessenger.setMockMessageHandler(
         'ch7',
         (ByteData message) async {
           final Map<dynamic, dynamic> methodCall = jsonMessage.decodeMessage(message);
@@ -60,7 +54,7 @@ void main() {
       expect(result, equals('hello world'));
     });
     test('can invoke list method and get result', () async {
-      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
+      defaultBinaryMessenger.setMockMessageHandler(
         'ch7',
         (ByteData message) async {
           final Map<dynamic, dynamic> methodCall = jsonMessage.decodeMessage(message);
@@ -76,7 +70,7 @@ void main() {
     });
 
     test('can invoke list method and get null result', () async {
-      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
+      defaultBinaryMessenger.setMockMessageHandler(
         'ch7',
         (ByteData message) async {
           final Map<dynamic, dynamic> methodCall = jsonMessage.decodeMessage(message);
@@ -92,7 +86,7 @@ void main() {
 
 
     test('can invoke map method and get result', () async {
-      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
+      defaultBinaryMessenger.setMockMessageHandler(
         'ch7',
         (ByteData message) async {
           final Map<dynamic, dynamic> methodCall = jsonMessage.decodeMessage(message);
@@ -108,7 +102,7 @@ void main() {
     });
 
     test('can invoke map method and get null result', () async {
-      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
+      defaultBinaryMessenger.setMockMessageHandler(
         'ch7',
         (ByteData message) async {
           final Map<dynamic, dynamic> methodCall = jsonMessage.decodeMessage(message);
@@ -123,7 +117,7 @@ void main() {
     });
 
     test('can invoke method and get error', () async {
-      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
+      defaultBinaryMessenger.setMockMessageHandler(
         'ch7',
         (ByteData message) async {
           return jsonMessage.encodeMessage(<dynamic>[
@@ -145,7 +139,7 @@ void main() {
       }
     });
     test('can invoke unimplemented method', () async {
-      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
+      defaultBinaryMessenger.setMockMessageHandler(
         'ch7',
         (ByteData message) async => null,
       );
@@ -163,7 +157,7 @@ void main() {
       channel.setMethodCallHandler(null);
       final ByteData call = jsonMethod.encodeMethodCall(const MethodCall('sayHello', 'hello'));
       ByteData envelope;
-      await ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage('ch7', call, (ByteData result) {
+      await defaultBinaryMessenger.handlePlatformMessage('ch7', call, (ByteData result) {
         envelope = result;
       });
       expect(envelope, isNull);
@@ -174,7 +168,7 @@ void main() {
       });
       final ByteData call = jsonMethod.encodeMethodCall(const MethodCall('sayHello', 'hello'));
       ByteData envelope;
-      await ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage('ch7', call, (ByteData result) {
+      await defaultBinaryMessenger.handlePlatformMessage('ch7', call, (ByteData result) {
         envelope = result;
       });
       expect(envelope, isNull);
@@ -183,7 +177,7 @@ void main() {
       channel.setMethodCallHandler((MethodCall call) async => '${call.arguments}, world');
       final ByteData call = jsonMethod.encodeMethodCall(const MethodCall('sayHello', 'hello'));
       ByteData envelope;
-      await ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage('ch7', call, (ByteData result) {
+      await defaultBinaryMessenger.handlePlatformMessage('ch7', call, (ByteData result) {
         envelope = result;
       });
       expect(jsonMethod.decodeEnvelope(envelope), equals('hello, world'));
@@ -194,7 +188,7 @@ void main() {
       });
       final ByteData call = jsonMethod.encodeMethodCall(const MethodCall('sayHello', 'hello'));
       ByteData envelope;
-      await ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage('ch7', call, (ByteData result) {
+      await defaultBinaryMessenger.handlePlatformMessage('ch7', call, (ByteData result) {
         envelope = result;
       });
       try {
@@ -213,7 +207,7 @@ void main() {
       });
       final ByteData call = jsonMethod.encodeMethodCall(const MethodCall('sayHello', 'hello'));
       ByteData envelope;
-      await ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage('ch7', call, (ByteData result) {
+      await defaultBinaryMessenger.handlePlatformMessage('ch7', call, (ByteData result) {
         envelope = result;
       });
       try {
@@ -232,7 +226,7 @@ void main() {
     const MethodCodec jsonMethod = JSONMethodCodec();
     const EventChannel channel = EventChannel('ch', jsonMethod);
     void emitEvent(dynamic event) {
-      ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage(
+      defaultBinaryMessenger.handlePlatformMessage(
         'ch',
         event,
             (ByteData reply) { },
@@ -240,7 +234,7 @@ void main() {
     }
     test('can receive event stream', () async {
       bool canceled = false;
-      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
+      defaultBinaryMessenger.setMockMessageHandler(
         'ch',
         (ByteData message) async {
           final Map<dynamic, dynamic> methodCall = jsonMessage.decodeMessage(message);
@@ -264,7 +258,7 @@ void main() {
       expect(canceled, isTrue);
     });
     test('can receive error event', () async {
-      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
+      defaultBinaryMessenger.setMockMessageHandler(
         'ch',
         (ByteData message) async {
           final Map<dynamic, dynamic> methodCall = jsonMessage.decodeMessage(message);
