+++ packages/flutter/test/services/system_sound_test.dart
@@ -6,10 +6,6 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  setUp(() {
-    TestWidgetsFlutterBinding.ensureInitialized();
-  });
-
   test('System sound control test', () async {
     final List<MethodCall> log = <MethodCall>[];
